Seperated Review Snapshot from Reviews/index file, to seperate component
