+++ styles/RevStats.module.css
@@ -1,27 +1,24 @@
-.container{
-  width: min(100%,17rem) ;
- margin-right: 2rem;
+.container {
+  width: min(100%, 17rem);
+  margin-right: 2rem;
 }
-
-.meterContainer{
-    display: flex;
-    flex-direction: row;
-    flex-wrap: nowrap;
-    justify-content: space-between;
+.globalRating{
+  color: #565959;
+  font-size: .85rem;
 }
-.meterContainer:hover .meter{
-    box-shadow: inset 0 0 0 1px  #de7921;
-    background-color: rgba(245, 150, 29, .15);
+.meterContainer:hover .meter {
+  box-shadow: inset 0 0 0 1px #de7921;
+  background-color: rgba(245, 150, 29, 0.15);
 }
-.hoverable:hover{
- text-decoration: underline;
+.hoverable:hover {
+  text-decoration: underline;
 }
-.zero{
-    color:#565959;
-    cursor: default;
+.zero {
+  color: #565959;
+  cursor: default;
 }
-.zero:hover{
-    color:#565959;
+.zero:hover {
+  color: #565959;
 }
 .meter {
   width: 10rem;
@@ -30,54 +27,54 @@
   box-shadow: inset 0 0 0 1px #e3e6e6;
   background-color: #f0f2f2;
   border-radius: 0.3rem;
-  margin-bottom: .1rem;
+  margin-left: 1rem;
+  margin-right: 1rem;
+  margin-bottom: .75rem;
 }
 .meter.fill {
-    margin: 0;
+  margin: 0;
   background-color: #f5961d !important;
   box-shadow: inset -2px 0 0 -1px #f5961d, inset 0 0 0 1px #de7921;
 }
-.showDropDown{
+.showDropDown {
   background: none;
   color: #007185;
   border: none;
   cursor: pointer;
   padding: 0;
-font-size: .8rem;
-
+  font-size: 0.8rem;
 }
-.showDropDown:hover{
+.showDropDown:hover {
   color: orange;
-text-decoration: underline;
+  text-decoration: underline;
 }
 
-.description{
-font-size: .9rem;
-line-height: 1.2rem;
+.description {
+  font-size: 0.9rem;
+  line-height: 1.2rem;
+  margin-top: 1rem;
 }
-.button{
+.button {
   width: 100%;
-  padding: .5rem 0;
-  border-radius: .7rem;
-  border: 1px solid rgba(15,17,17,.2);
+  padding: 0.5rem 0;
+  border-radius: 0.7rem;
+  border: 1px solid rgba(15, 17, 17, 0.2);
   background-color: transparent;
   cursor: pointer;
-  font-size: .75rem;
+  font-size: 0.75rem;
 }
 .button:hover {
-  background-color: rgba(244,244,244,.4);
-
+  background-color: rgba(244, 244, 244, 0.4);
 }
-.createReview{
-  border-bottom:1px solid rgba(15,17,17,.2) ;
+.createReview {
+  border-bottom: 1px solid rgba(15, 17, 17, 0.2);
   padding: 2rem 0;
   margin: 2rem 0;
-  border-top:1px solid rgba(15,17,17,.2) ;
-  
+  border-top: 1px solid rgba(15, 17, 17, 0.2);
 }
 .createReview > h3 {
   margin: 0;
 }
-.createReview > p{
-  font-size: .75rem;
+.createReview > p {
+  font-size: 0.75rem;
 }
